# api/internal/server/routes.gen.go
//go:build go1.22

// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package server

import (
	"bytes"
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for TodoState.
const (
	TodoStateComplete   TodoState = "complete"
	TodoStateIncomplete TodoState = "incomplete"
)

// Valid indicates whether the value is a known member of the TodoState enum.
func (e TodoState) Valid() bool {
	switch e {
	case TodoStateComplete:
		return true
	case TodoStateIncomplete:
		return true
	default:
		return false
	}
}

// Defines values for GetTodosParamsState.
const (
	GetTodosParamsStateComplete   GetTodosParamsState = "complete"
	GetTodosParamsStateIncomplete GetTodosParamsState = "incomplete"
)

// Valid indicates whether the value is a known member of the GetTodosParamsState enum.
func (e GetTodosParamsState) Valid() bool {
	switch e {
	case GetTodosParamsStateComplete:
		return true
	case GetTodosParamsStateIncomplete:
		return true
	default:
		return false
	}
}

// ErrorMessage defines model for ErrorMessage.
type ErrorMessage struct {
	// Message Reason for the error response
	Message string `json:"message"`
}

// FieldError defines model for FieldError.
type FieldError struct {
	// Code A unique code for the error type.
	Code string `json:"code"`

	// Field The field the error is associated with.
	Field string `json:"field"`

	// Title A human-readable summary of the error.
	Title string `json:"title"`
}

// NewTodo defines model for NewTodo.
type NewTodo struct {
	// Title Title of the todo
	Title string `json:"title"`
}

// ProblemDetails defines model for ProblemDetails.
type ProblemDetails struct {
	// Detail A human-readable explanation specific to this occurrence of the problem.
	Detail *string `json:"detail,omitempty"`

	// FieldErrors Errors associated with specific fields of the request.
	FieldErrors *[]FieldError `json:"fieldErrors,omitempty"`

	// Title A short, human-readable summary of the problem type.
	Title *string `json:"title,omitempty"`

	// Type A URI reference that identifies the problem type.
	Type *string `json:"type,omitempty"`
}

// Todo defines model for Todo.
type Todo struct {
	// CompletedAt Instant when the todo was marked as complete
	CompletedAt *time.Time `json:"completedAt,omitempty"`

	// Id A unique identifier for the todo
	Id int `json:"id"`

	// Title Title of the todo
	Title string `json:"title"`
}

// TodoCollection defines model for TodoCollection.
type TodoCollection struct {
	Items []Todo `json:"items"`
}

// TodoState A possible state for a todo item
type TodoState string

// BadRequest defines model for BadRequest.
type BadRequest = ProblemDetails

// TodoNotFound defines model for TodoNotFound.
type TodoNotFound = ErrorMessage

// GetTodosParams defines parameters for GetTodos.
type GetTodosParams struct {
	State *GetTodosParamsState `form:"state,omitempty" json:"state,omitempty"`
}

// GetTodosParamsState defines parameters for GetTodos.
type GetTodosParamsState string

// PostTodosJSONRequestBody defines body for PostTodos for application/json ContentType.
type PostTodosJSONRequestBody = NewTodo

// PutTodosTodoIDStateTextRequestBody defines body for PutTodosTodoIDState for text/plain ContentType.
type PutTodosTodoIDStateTextRequestBody = TodoState

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /todos)
	GetTodos(w http.ResponseWriter, r *http.Request, params GetTodosParams)

	// (POST /todos)
	PostTodos(w http.ResponseWriter, r *http.Request)

	// (PUT /todos/{todoID}/state)
	PutTodosTodoIDState(w http.ResponseWriter, r *http.Request, todoID int)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetTodos operation middleware
func (siw *ServerInterfaceWrapper) GetTodos(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTodosParams

	// ------------- Optional query parameter "state" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "state", r.URL.Query(), &params.State, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "state"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "state", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTodos(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostTodos operation middleware
func (siw *ServerInterfaceWrapper) PostTodos(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostTodos(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PutTodosTodoIDState operation middleware
func (siw *ServerInterfaceWrapper) PutTodosTodoIDState(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "todoID" -------------
	var todoID int

	err = runtime.BindStyledParameterWithOptions("simple", "todoID", r.PathValue("todoID"), &todoID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "todoID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PutTodosTodoIDState(w, r, todoID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/todos", wrapper.GetTodos)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/todos", wrapper.PostTodos)
	m.HandleFunc(http.MethodPut+" "+options.BaseURL+"/todos/{todoID}/state", wrapper.PutTodosTodoIDState)

	return m
}

type BadRequestApplicationProblemPlusJSONResponse ProblemDetails

type TodoNotFoundJSONResponse ErrorMessage

type GetTodosRequestObject struct {
	Params GetTodosParams
}

type GetTodosResponseObject interface {
	VisitGetTodosResponse(w http.ResponseWriter) error
}

type GetTodos200JSONResponse TodoCollection

func (response GetTodos200JSONResponse) VisitGetTodosResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PostTodosRequestObject struct {
	Body *PostTodosJSONRequestBody
}

type PostTodosResponseObject interface {
	VisitPostTodosResponse(w http.ResponseWriter) error
}

type PostTodos201JSONResponse Todo

func (response PostTodos201JSONResponse) VisitPostTodosResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type PostTodos400ApplicationProblemPlusJSONResponse struct {
	BadRequestApplicationProblemPlusJSONResponse
}

func (response PostTodos400ApplicationProblemPlusJSONResponse) VisitPostTodosResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PutTodosTodoIDStateRequestObject struct {
	TodoID int `json:"todoID"`
	Body   *PutTodosTodoIDStateTextRequestBody
}

type PutTodosTodoIDStateResponseObject interface {
	VisitPutTodosTodoIDStateResponse(w http.ResponseWriter) error
}

type PutTodosTodoIDState200JSONResponse Todo

func (response PutTodosTodoIDState200JSONResponse) VisitPutTodosTodoIDStateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PutTodosTodoIDState400ApplicationProblemPlusJSONResponse struct {
	BadRequestApplicationProblemPlusJSONResponse
}

func (response PutTodosTodoIDState400ApplicationProblemPlusJSONResponse) VisitPutTodosTodoIDStateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PutTodosTodoIDState404JSONResponse struct{ TodoNotFoundJSONResponse }

func (response PutTodosTodoIDState404JSONResponse) VisitPutTodosTodoIDStateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /todos)
	GetTodos(ctx context.Context, request GetTodosRequestObject) (GetTodosResponseObject, error)

	// (POST /todos)
	PostTodos(ctx context.Context, request PostTodosRequestObject) (PostTodosResponseObject, error)

	// (PUT /todos/{todoID}/state)
	PutTodosTodoIDState(ctx context.Context, request PutTodosTodoIDStateRequestObject) (PutTodosTodoIDStateResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetTodos operation middleware
func (sh *strictHandler) GetTodos(w http.ResponseWriter, r *http.Request, params GetTodosParams) {
	var request GetTodosRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetTodos(ctx, request.(GetTodosRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTodos")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetTodosResponseObject); ok {
		if err := validResponse.VisitGetTodosResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostTodos operation middleware
func (sh *strictHandler) PostTodos(w http.ResponseWriter, r *http.Request) {
	var request PostTodosRequestObject

	var body PostTodosJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		if !errors.Is(err, io.EOF) {
			sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
			return
		}
	} else {
		request.Body = &body
	}

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PostTodos(ctx, request.(PostTodosRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostTodos")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PostTodosResponseObject); ok {
		if err := validResponse.VisitPostTodosResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PutTodosTodoIDState operation middleware
func (sh *strictHandler) PutTodosTodoIDState(w http.ResponseWriter, r *http.Request, todoID int) {
	var request PutTodosTodoIDStateRequestObject

	request.TodoID = todoID

	data, err := io.ReadAll(r.Body)
	if err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't read body: %w", err))
		return
	}
	if len(data) > 0 {
		body := PutTodosTodoIDStateTextRequestBody(data)
		request.Body = &body
	}

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PutTodosTodoIDState(ctx, request.(PutTodosTodoIDStateRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PutTodosTodoIDState")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PutTodosTodoIDStateResponseObject); ok {
		if err := validResponse.VisitPutTodosTodoIDStateResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}