# openapi.yaml
openapi: "3.0.0"
info:
  version: "0.0.1"
  title: Todos

servers:
  - url: http://localhost:8000/api/v1

paths:
  /todos:
    get:
      parameters:
        - in: query
          name: state
          required: false
          schema:
            type: string
            enum:
              - complete
              - incomplete
            default: incomplete

      responses:
        "200":
          description: Todo collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoCollection"

    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTodo"
      responses:
        "201":
          description: Todo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"

  /todos/{todoID}/state:
    parameters:
      - name: todoID
        in: path
        required: true
        schema:
          type: integer
          description: ID of the todo to query
          example: 42

    put:
      requestBody:
        content:
          text/plain:
            schema:
              $ref: "#/components/schemas/TodoState"
      responses:
        "200":
          description: State updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/TodoNotFound"

components:
  schemas:
    # RFC 9457 compliant problem schema
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri-reference
          description: A URI reference that identifies the problem type.
          example: /probs/validation-error
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: Bad Request
        detail:
          type: string
          description: >
            A human-readable explanation specific to this occurrence of the
            problem.
          example: Some fields contain bad data.
        fieldErrors:
          type: array
          description: Errors associated with specific fields of the request.
          items:
            $ref: "#/components/schemas/FieldError"

    FieldError:
      type: object
      required:
        - field
        - code
        - title
      properties:
        field:
          type: string
          description: The field the error is associated with.
          example: username
        code:
          type: string
          description: A unique code for the error type.
          example: "not-unique"
        title:
          type: string
          description: A human-readable summary of the error.
          example: The username "jane" is already taken.

    ErrorMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Reason for the error response
          example: Something has gone terribly wrong.

    TodoCollection:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Todo"

    Todo:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: integer
          description: A unique identifier for the todo
          example: 4
        title:
          type: string
          description: Title of the todo
          example: Fold laundry
        completedAt:
          type: string
          format: date-time
          description: Instant when the todo was marked as complete
          example: 2026-05-28T16:32:28Z

    NewTodo:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the todo
          example: Fold laundry

    TodoState:
      type: string
      enum:
        - complete
        - incomplete
      description: A possible state for a todo item

  responses:
    BadRequest:
      description: The request was invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    TodoNotFound:
      description: No todo found with the given ID
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            message: No todo with ID 7.